Ups, error

+++ src/settlers/Utilities.java
@@ -107,7 +107,7 @@ public class Utilities {
 		return r;
 	}
 	
-	public static float RandomInt(int start, int end)
+	public static int RandomInt(int start, int end)
 	{
 		Random random = new Random();
 		float r = Math.abs(random.nextInt());
